+++ packages/ckeditor5-dev-ci/lib/notifytravisstatus.js
@@ -37,19 +37,12 @@ module.exports = async function notifyTravisStatus( options ) {
 	const commitDetails = await getCommitDetails( commitUrl, options.githubToken );
 
 	const [ buildRepoOwner, buildRepoName ] = options.repositorySlug.split( '/' );
-	const execTime = getExecutionTime( options.endTime, options.startTime );
 
 	const slackAccount = members[ commitDetails.author ];
 	const shortCommit = commitUrl.split( '/' ).pop().substring( 0, 7 );
 	const repoMatch = commitUrl.match( REPOSITORY_REGEXP );
 
-	const slack = slackNotify( options.slackWebhookUrl );
-
-	slack.onError = err => {
-		console.log( 'API error occurred:', err );
-	};
-
-	return slack.send( {
+	const messageData = {
 		username: 'Travis CI',
 		text: getNotifierMessage( {
 			slackAccount,
@@ -82,7 +75,7 @@ module.exports = async function notifyTravisStatus( options ) {
 					},
 					{
 						title: 'Build time',
-						value: `${ execTime.mins } min ${ execTime.secs } sec`,
+						value: getExecutionTime( options.endTime, options.startTime ),
 						short: true
 					},
 					{
@@ -93,27 +86,42 @@ module.exports = async function notifyTravisStatus( options ) {
 				]
 			}
 		]
-	} );
+	};
+
+	return slackNotify( options.slackWebhookUrl )
+		.send( messageData )
+		.catch( err => console.log( 'API error occurred:', err ) );
 };
 
 /**
- * Returns an object that compares two dates.
+ * Returns string representing amount of time passed between two timestamps.
  *
  * @param {Number} endTime
  * @param {Number} startTime
- * @returns {Object}
+ * @returns {String}
  */
 function getExecutionTime( endTime, startTime ) {
-	const execTime = {
-		ms: endTime - startTime
-	};
+	const totalMs = ( endTime - startTime ) * 1000;
+	const date = new Date( totalMs );
+	const hours = date.getUTCHours();
+	const minutes = date.getUTCMinutes();
+	const seconds = date.getUTCSeconds();
 
-	execTime.days = Math.floor( execTime.ms / 86400 );
-	execTime.hours = Math.floor( ( execTime.ms - 86400 * execTime.days ) / 3600 );
-	execTime.mins = Math.floor( ( ( execTime.ms - 86400 * execTime.days ) - 3600 * execTime.hours ) / 60 );
-	execTime.secs = ( ( execTime.ms - 86400 * execTime.days ) - 3600 * execTime.hours ) - 60 * execTime.mins;
+	const stringParts = [];
 
-	return execTime;
+	if ( hours ) {
+		stringParts.push( `${ hours } hr.` );
+	}
+
+	if ( minutes ) {
+		stringParts.push( `${ minutes } min.` );
+	}
+
+	if ( seconds ) {
+		stringParts.push( `${ seconds } sec.` );
+	}
+
+	return stringParts.join( ' ' );
 }
 
 /**