+++ packages/ckeditor5-dev-tests/lib/utils/automated-tests/getwebpackconfig.js
@@ -28,6 +28,7 @@ module.exports = function getWebpackConfigForAutomatedTests( options ) {
 		plugins: [
 			new webpack.DefinePlugin( definitions ),
 			new webpack.ProvidePlugin( {
+				Buffer: [ 'buffer', 'Buffer' ],
 				process: 'process/browser'
 			} )
 		],