+++ packages/typedoc-plugins/tests/event-inheritance-fixer/fixtures/class-mixin-foobar.ts
@@ -1,0 +1,13 @@
+/**
+ * @license Copyright (c) 2003-2023, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md.
+ */
+
+import { ClassMixinFoo } from './class-mixin-foo';
+
+/**
+ * @module fixtures/class-mixin-foobar
+ */
+
+export class ClassMixinFooBar extends ClassMixinFoo {
+}